add SoundBuilder and change Controller to u64

+++ src/audio/host_ffi.rs
@@ -3,7 +3,7 @@ unsafe extern "C" {
     pub(crate) fn boppo_play_sound_instruction(si_ptr: *const u8, si_length: usize) -> i32;
 
     pub(crate) fn boppo_set_controller_parameter(
-        controller_id: i32,
+        controller_id: u64,
         parameter: i32,
         value: f32,
     ) -> i32;
